# api/swagger.yaml
---
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Exercises API"
tags:
- name: "Exercise"
  description: "Endpoints for Zolara's exercises"
paths:
  /userExercise:
    post:
      tags:
      - "Exercise"
      summary: "Creates Personalized exercises for the User"
      operationId: "createUserExercises"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "Successful creation of exercises for the current user"
          schema:
            $ref: "#/definitions/UserExercises"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "404":
          description: "Not Found"
      security:
      - USER: []
    put:
      tags:
      - "Exercise"
      summary: "Update results of the exercises"
      operationId: "updateUserExercises"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "Successful update of exercises for the current user"
          schema:
            $ref: "#/definitions/ExercisesAnswers"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "404":
          description: "Not Found"
securityDefinitions:
  USER:
    type: "apiKey"
    name: "X-Csrf-Token"
    in: "header"
definitions:
  UserExercises:
    type: "object"
    required:
    - "flashCards"
    - "gapFills"
    properties:
      gapFills:
        type: "array"
        items:
          $ref: "#/definitions/GapFillList"
      flashCards:
        type: "array"
        items:
          $ref: "#/definitions/FlashCardList"
    example:
      flashCards:
      - answer:
        - answer: "answer"
          answerOrder: 6
          isCorrect: true
        - answer: "answer"
          answerOrder: 6
          isCorrect: true
        imageUrl: "imageUrl"
        id: 1
      - answer:
        - answer: "answer"
          answerOrder: 6
          isCorrect: true
        - answer: "answer"
          answerOrder: 6
          isCorrect: true
        imageUrl: "imageUrl"
        id: 1
      gapFills:
      - answer:
        - answer: "answer"
          answerOrder: 6
          isCorrect: true
        - answer: "answer"
          answerOrder: 6
          isCorrect: true
        id: 0
        text: "text"
      - answer:
        - answer: "answer"
          answerOrder: 6
          isCorrect: true
        - answer: "answer"
          answerOrder: 6
          isCorrect: true
        id: 0
        text: "text"
  FlashCardList:
    type: "object"
    required:
    - "answer"
    - "id"
    - "imageUrl"
    properties:
      id:
        type: "integer"
        format: "int64"
      imageUrl:
        type: "string"
      answer:
        type: "array"
        items:
          $ref: "#/definitions/Answers"
    example:
      answer:
      - answer: "answer"
        answerOrder: 6
        isCorrect: true
      - answer: "answer"
        answerOrder: 6
        isCorrect: true
      imageUrl: "imageUrl"
      id: 1
  GapFillList:
    type: "object"
    required:
    - "answer"
    - "id"
    - "text"
    properties:
      id:
        type: "integer"
        format: "int64"
      text:
        type: "string"
      answer:
        type: "array"
        items:
          $ref: "#/definitions/Answers"
    example:
      answer:
      - answer: "answer"
        answerOrder: 6
        isCorrect: true
      - answer: "answer"
        answerOrder: 6
        isCorrect: true
      id: 0
      text: "text"
  Answers:
    type: "object"
    required:
    - "answer"
    - "isCorrect"
    properties:
      answer:
        type: "string"
      isCorrect:
        type: "boolean"
      answerOrder:
        type: "integer"
        format: "int64"
    example:
      answer: "answer"
      answerOrder: 6
      isCorrect: true
  ExerciseAnswer:
    type: "object"
    required:
    - "completed"
    - "id"
    - "missed"
    properties:
      id:
        type: "integer"
        format: "int64"
      completed:
        type: "boolean"
      missed:
        type: "boolean"
    example:
      missed: true
      id: 0
      completed: true
  ExercisesAnswers:
    type: "object"
    required:
    - "answers"
    properties:
      answers:
        type: "array"
        items:
          $ref: "#/definitions/ExerciseAnswer"
    example:
      answers:
      - missed: true
        id: 0
        completed: true
      - missed: true
        id: 0
        completed: true