# models/model_answers.go
/*
 * Exercises API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package models

type Answers struct {

	Answer string `json:"answer"`

	IsCorrect bool `json:"isCorrect"`

	AnswerOrder int64 `json:"answerOrder,omitempty"`
}