# models/model_exercise_answer.go
/*
 * Exercises API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package models

type ExerciseAnswer struct {

	Id int64 `json:"id"`

	Completed bool `json:"completed"`

	Missed bool `json:"missed"`
}