# models/model_exercises_answers.go
/*
 * Exercises API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package models

type ExercisesAnswers struct {

	Answers []ExerciseAnswer `json:"answers"`
}