# models/model_flash_card_list.go
/*
 * Exercises API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package models

type FlashCardList struct {

	Id int64 `json:"id"`

	ImageUrl string `json:"imageUrl"`

	Answer []Answers `json:"answer"`
}