# models/model_gap_fill_list.go
/*
 * Exercises API
 *
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package models

type GapFillList struct {

	Id int64 `json:"id"`

	Text string `json:"text"`

	Answer []Answers `json:"answer"`
}